Use a server ring instead of a cycling iterator
Guava's cycling iterator does not seem to be threadsafe, as it replaces the underlying iterator between
calling `hasNext()` and `next()`, which can cause a `ConcurrentModificationException`.

Taking into account the research and design suggestions from @alkiskal and @jasonculverhouse, here
is a reworked version of #313 for #311 which simply keeps a next index pointer, and a list of servers
that acts as a ring.

Closes #313
Closes #311