Update default Gson to support long values.
Background:
When a Search object is prepared for an HttpRequest the payload of a
query remains untouched except when sorting. When sorting the Search
object deserializes the query to a Map using Gson. The Gson by default
serializes Numbers to Doubles. If you happened to use a Long of '1234'
it will be converted to '1234.0'. Upon serialization a query's search
value will not reflect the original setting.

Code Changes:
- Modified the default Gson to support long values by adding a custom
serializer to the default Gson.
- Added a Unit Test to verify longs remain longs.

Resolves: #335