Make adding metadata fields to JestResult optional
JestResult#getSourceAsObjectList() and JestResult#getSourceAsObject() used to
add the metadata fields "es_metadata_id" and "es_metadata_version" unconditionally,
which might be the desired for all use cases.

This change set adds a boolean parameter to these methods which controls whether
the metadata fields should be added to the result.